Support decimal catalog numbers for personal picks outside the 1001 list
Personal album picks (not part of the book) get numbered like "663.1",
inserted right after book entry 663 in Medium posts. Scraper regex,
DB insert/lookup, and frontend sort now handle the decimal, and the
dashboard progress ring excludes non-integer numbers so completion%
still reflects actual book progress.

+++ db.py
@@ -53,7 +53,7 @@ def find_album_id(conn: sqlite3.Connection, number, artist: str, album: str) -> 
     always merged on: (catalog_number, artist, album)."""
     row = conn.execute(
         "SELECT id FROM albums WHERE catalog_number = ? AND artist = ? AND album = ?",
-        (int(number), artist, album),
+        (float(number), artist, album),
     ).fetchone()
     return row["id"] if row else None
 
@@ -85,7 +85,7 @@ def insert_album(conn: sqlite3.Connection, e: dict) -> int:
         ) VALUES (?, ?, ?, ?, ?, NULL, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
         """,
         (
-            int(e["number"]), e["artist"], e["album"], int(e["year"]), e["text"],
+            float(e["number"]), e["artist"], e["album"], int(e["year"]), e["text"],
             e.get("medium_post_url"),
             spotify.get("spotify_url"), spotify.get("spotify_embed_url"),
             spotify.get("cover_art_url"), spotify.get("matched_artist_name"),
